requirements.txt
langchain
langchain-groq
pydantic
python-dotenv
langchain_core
fastapi
uvicorn
requests
pydantic
streamlit